Move SelectionMode into its own file
egui's CONTRIBUTING.md asks for each type in its own file unless trivial.
Move the `SelectionMode` enum and its impl out of `text_cursor_state.rs`
into a new `selection_mode.rs`, and make the `select_word_at`,
`select_line_at` and `range_bounds` helpers `pub(crate)` so the new
module can use them. The `from_click_count` test moves along with it.

No behavior change.
